Providing feedback
So, this design worked perfectly almost right out of the box. I've attached the constraints file with the caps fixed for your buttons (which was my fault with the instructions). That is all I had to do to get it working.

I'll go ahead and fill you and Hannah in for 90% on this assignment, pending a demo this Tuesday. Excellent work.